Approach 1.  Fixes constant reboot for UI7 but probably breaks UI5
Not sure if this is supposed to work on UI5.
When adding BinaryLight or DimmableLight modules (A1,A2,etc), Mochad causes Vera to repeatedly crash and reboot on UI7.
This is because in UI7 the 'schema' parameter in the call to  'luup.chdev.append' used in function 'add_children' must either be blank or correct. (MCV has an interesting approach to dealing with errors)
Credit Bandanafish in  Reply #1 in the thread at: http://forum.micasaverde.com/index.php/topic,32747.msg241189.html#msg241189
MCV, of course, decided to change these values for UI7.  This is why I think this code will probably break UI5, so maybe it's better to just change the value for 'schema' to "".